# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Inventory Management API Documentation",
        "description": "Dokumentasi API untuk Sistem Inventaris Tahap 1-4",
        "contact": {
            "email": "admin@inventaris.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "API Server Utama"
        }
    ],
    "paths": {
...
        "/api/proxy/areas": {
            "get": {
                "tags": [
                    "Logistics Gateway"
                ],
                "summary": "Cari Area Biteship (Kecamatan/Kodepos)",
                "description": "Mencari area di Indonesia menggunakan Biteship API berdasarkan nama kecamatan atau kode pos (min 3 karakter). Memerlukan hak akses Admin atau Staff.",
                "operationId": "searchAreas",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Nama kecamatan atau kode pos",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "Mampang",
                            "minLength": 3
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil melakukan pencarian area",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "string",
                                                        "example": "ID-12345"
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Mampang Prapatan"
                                                    },
                                                    "postal_code": {
                                                        "type": "integer",
                                                        "example": 12790
                                                    },
                                                    "country": {
                                                        "type": "string",
                                                        "example": "ID"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Token JWT tidak valid atau kadaluwarsa"
                    },
                    "500": {
                        "description": "Gagal mencari area karena kendala gateway api"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/proxy/currency-rates": {
            "get": {
                "tags": [
                    "Finance Gateway"
                ],
                "summary": "Ambil kurs mata uang asing (USD ke IDR)",
                "description": "Mendapatkan kurs konversi mata uang asing real-time (USD, SGD, CNY ke IDR) untuk kebutuhan impor barang/komponen. Memerlukan hak akses Admin atau Staff.",
                "operationId": "getCurrencyRates",
                "responses": {
                    "200": {
                        "description": "Kurs berhasil diambil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Kurs berhasil diambil"
                                        },
                                        "data": {
                                            "properties": {
                                                "base": {
                                                    "type": "string",
                                                    "example": "USD"
                                                },
                                                "rates": {
                                                    "properties": {
                                                        "IDR": {
                                                            "type": "number",
                                                            "example": 16250.5
                                                        },
                                                        "SGD": {
                                                            "type": "number",
                                                            "example": 1.35
                                                        },
                                                        "CNY": {
                                                            "type": "number",
                                                            "example": 7.24
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "last_updated": {
                                                    "type": "string",
                                                    "example": "Mon, 08 Jun 2026 00:00:00 +0000"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Terjadi kesalahan pada server currency gateway"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}